fix: reposition on screen out-of-bounds (#16)
This adds an example app to run the widget locally.
Further, check after the layout run, if the size of the
widget would be outside the screen. If it is, reposition
the match the screen. Also, if the content is wider/taller
than the screen, it maxes to 0 on the respective sides.